+++ concordium-consensus/src/Concordium/Scheduler/EnvironmentImplementation.hs
@@ -119,6 +119,10 @@ instance (MonadReader ContextState m,
   {-# INLINE getStateAccount #-}
   getStateAccount !addr = lift . flip bsoGetAccount addr =<< use schedulerBlockState
 
+  {-# INLINE getExchangeRates #-}
+  getExchangeRates = lift . bsoGetExchangeRates =<< use schedulerBlockState
+
+
 instance (SS state ~ UpdatableBlockState m,
           HasSchedulerState state,
           MonadState state m,
@@ -157,7 +161,7 @@ instance (SS state ~ UpdatableBlockState m,
   {-# INLINE energyToCcd #-}
   energyToCcd v =  do
     s <- use schedulerBlockState
-    rate <- lift (bsoGetEnergyRate s)
+    rate <- lift $ _erEnergyRate <$> bsoGetExchangeRates s
     return (computeCost rate v)
   {-# INLINE getMaxBlockEnergy #-}
   getMaxBlockEnergy = do
@@ -250,8 +254,8 @@ instance (MonadReader ContextState m,
   {-# INLINE energyToGtu #-}
   energyToGtu v = do
     s <- use schedulerBlockState
-    rate <- lift (bsoGetEnergyRate s)
-    return (computeCost rate v)
+    rate <- lift $ _erEnergyRate <$> bsoGetExchangeRates s
+    return $! computeCost rate v
 
   {-# INLINE notifyExecutionCost #-}
   notifyExecutionCost !amnt = schedulerExecutionCosts += amnt